Support contaminants
Contaminants are basis functions with angular part corresponding
to a lower l, and multiplied by (r^2)^p; formally, they belong
to the shell with a higher l (= l+2p), and therefore use the
latter's exponents and coefficients. In HDF5 files, contaminants
are found as functions in a particular center, with a lower l
coming after functions with a higher l, since normally the functions
are ordered by l.