Add a cache for primitive radial parts
In general contraction, different basis functions use the same
radial primitives again and again. This add a (hopefully small)
in-memory cache of primitives, so they can be reused in later
functions. Note that functions of the same l and different m
already share the same radial part, and this is not recomputed at all.