Normalize basis function coefficients
Until now, Pegamoid was assuming that the primitive coefficients
(i.e., the basis functions) are normalized, and that the MO
coefficients refer to the input basis as given.

This is not necessarily the case. The MO coefficients refer to
the normalized basis functions, but the input primitive coefficients
may not be normalized. Therefore, a prenormalization step is
introduced.

Some basis sets provide already normalized coefficients, apparently
all the sample calculations, so this bug has probably not had any
noticeable effect in many cases.